+++ packages/cli/src/ui/components/messages/ShellToolMessage.tsx
@@ -6,52 +6,29 @@
 
 import React from 'react';
 import { Box, Text, type DOMElement } from 'ink';
-import type { IndividualToolCallDisplay } from '../../types.js';
 import { ToolCallStatus } from '../../types.js';
-import { DiffRenderer } from './DiffRenderer.js';
-import { MarkdownDisplay } from '../../utils/MarkdownDisplay.js';
-import { AnsiOutputText } from '../AnsiOutput.js';
-import { GeminiRespondingSpinner } from '../GeminiRespondingSpinner.js';
-import { MaxSizedBox } from '../shared/MaxSizedBox.js';
 import { ShellInputPrompt } from '../ShellInputPrompt.js';
 import { StickyHeader } from '../StickyHeader.js';
-import {
-  SHELL_COMMAND_NAME,
-  SHELL_NAME,
-  TOOL_STATUS,
-} from '../../constants.js';
+import { SHELL_COMMAND_NAME, SHELL_NAME } from '../../constants.js';
 import { theme } from '../../semantic-colors.js';
-import type { AnsiOutput, Config } from '@google/gemini-cli-core';
 import { SHELL_TOOL_NAME } from '@google/gemini-cli-core';
-import { useUIState } from '../../contexts/UIStateContext.js';
-import { useAlternateBuffer } from '../../hooks/useAlternateBuffer.js';
 import { useUIActions } from '../../contexts/UIActionsContext.js';
 import { useMouseClick } from '../../hooks/useMouseClick.js';
+import { ToolResultDisplay } from './ToolResultDisplay.js';
+import {
+  ToolStatusIndicator,
+  ToolInfo,
+  TrailingIndicator,
+  STATUS_INDICATOR_WIDTH,
+} from './ToolShared.js';
+import type { ToolMessageProps } from './ToolMessage.js';
 
-const STATIC_HEIGHT = 1;
-const RESERVED_LINE_COUNT = 5; // for tool name, status, padding etc.
-const STATUS_INDICATOR_WIDTH = 3;
-const MIN_LINES_SHOWN = 2; // show at least this many lines
-
-// Large threshold to ensure we don't cause performance issues for very large
-// outputs that will get truncated further MaxSizedBox anyway.
-const MAXIMUM_RESULT_DISPLAY_CHARACTERS = 1000000;
-export type TextEmphasis = 'high' | 'medium' | 'low';
-
-export interface ToolMessageProps extends IndividualToolCallDisplay {
-  availableTerminalHeight?: number;
-  terminalWidth: number;
-  emphasis?: TextEmphasis;
-  renderOutputAsMarkdown?: boolean;
+interface ShellToolMessageProps extends ToolMessageProps {
   activeShellPtyId?: number | null;
   embeddedShellFocused?: boolean;
-  isFirst: boolean;
-  borderColor: string;
-  borderDimColor: boolean;
-  config?: Config;
 }
 
-export const ShellToolMessage: React.FC<ToolMessageProps> = ({
+export const ShellToolMessage: React.FC<ShellToolMessageProps> = ({
   name,
   description,
   resultDisplay,
@@ -68,11 +45,9 @@ export const ShellToolMessage: React.FC<ToolMessageProps> = ({
   borderColor,
   borderDimColor,
 }) => {
-  const { renderMarkdown } = useUIState();
-  const isAlternateBuffer = useAlternateBuffer();
   const isThisShellFocused =
     (name === SHELL_COMMAND_NAME ||
-      name === 'Shell' ||
+      name === SHELL_NAME ||
       name === SHELL_TOOL_NAME) &&
     status === ToolCallStatus.Executing &&
     ptyId === activeShellPtyId &&
@@ -83,7 +58,7 @@ export const ShellToolMessage: React.FC<ToolMessageProps> = ({
   // The shell is focusable if it's the shell command, it's executing, and the interactive shell is enabled.
   const isThisShellFocusable =
     (name === SHELL_COMMAND_NAME ||
-      name === 'Shell' ||
+      name === SHELL_NAME ||
       name === SHELL_TOOL_NAME) &&
     status === ToolCallStatus.Executing &&
     config?.getEnableInteractiveShell();
@@ -141,95 +116,6 @@ export const ShellToolMessage: React.FC<ToolMessageProps> = ({
   const shouldShowFocusHint =
     isThisShellFocusable && (showFocusHint || userHasFocused);
 
-  const availableHeight = availableTerminalHeight
-    ? Math.max(
-        availableTerminalHeight - STATIC_HEIGHT - RESERVED_LINE_COUNT,
-        MIN_LINES_SHOWN + 1, // enforce minimum lines shown
-      )
-    : undefined;
-
-  // Long tool call response in MarkdownDisplay doesn't respect availableTerminalHeight properly,
-  // so if we aren't using alternate buffer mode, we're forcing it to not render as markdown when the response is too long, it will fallback
-  // to render as plain text, which is contained within the terminal using MaxSizedBox
-  if (availableHeight && !isAlternateBuffer) {
-    renderOutputAsMarkdown = false;
-  }
-  const combinedPaddingAndBorderWidth = 4;
-  const childWidth = terminalWidth - combinedPaddingAndBorderWidth;
-
-  const truncatedResultDisplay = React.useMemo(() => {
-    if (typeof resultDisplay === 'string') {
-      if (resultDisplay.length > MAXIMUM_RESULT_DISPLAY_CHARACTERS) {
-        return '...' + resultDisplay.slice(-MAXIMUM_RESULT_DISPLAY_CHARACTERS);
-      }
-    }
-    return resultDisplay;
-  }, [resultDisplay]);
-
-  const renderedResult = React.useMemo(() => {
-    if (!truncatedResultDisplay) return null;
-
-    return (
-      <Box width={childWidth} flexDirection="column">
-        <Box flexDirection="column">
-          {typeof truncatedResultDisplay === 'string' &&
-          renderOutputAsMarkdown ? (
-            <Box flexDirection="column">
-              <MarkdownDisplay
-                text={truncatedResultDisplay}
-                terminalWidth={childWidth}
-                renderMarkdown={renderMarkdown}
-                isPending={false}
-              />
-            </Box>
-          ) : typeof truncatedResultDisplay === 'string' &&
-            !renderOutputAsMarkdown ? (
-            isAlternateBuffer ? (
-              <Box flexDirection="column" width={childWidth}>
-                <Text wrap="wrap" color={theme.text.primary}>
-                  {truncatedResultDisplay}
-                </Text>
-              </Box>
-            ) : (
-              <MaxSizedBox maxHeight={availableHeight} maxWidth={childWidth}>
-                <Box>
-                  <Text wrap="wrap" color={theme.text.primary}>
-                    {truncatedResultDisplay}
-                  </Text>
-                </Box>
-              </MaxSizedBox>
-            )
-          ) : typeof truncatedResultDisplay === 'object' &&
-            'fileDiff' in truncatedResultDisplay ? (
-            <DiffRenderer
-              diffContent={truncatedResultDisplay.fileDiff}
-              filename={truncatedResultDisplay.fileName}
-              availableTerminalHeight={availableHeight}
-              terminalWidth={childWidth}
-            />
-          ) : typeof truncatedResultDisplay === 'object' &&
-            'todos' in truncatedResultDisplay ? (
-            // display nothing, as the TodoTray will handle rendering todos
-            <></>
-          ) : (
-            <AnsiOutputText
-              data={truncatedResultDisplay as AnsiOutput}
-              availableTerminalHeight={availableHeight}
-              width={childWidth}
-            />
-          )}
-        </Box>
-      </Box>
-    );
-  }, [
-    truncatedResultDisplay,
-    renderOutputAsMarkdown,
-    childWidth,
-    renderMarkdown,
-    isAlternateBuffer,
-    availableHeight,
-  ]);
-
   return (
     <Box ref={containerRef} flexDirection="column" width={terminalWidth}>
       <StickyHeader
@@ -266,7 +152,12 @@ export const ShellToolMessage: React.FC<ToolMessageProps> = ({
         paddingX={1}
         flexDirection="column"
       >
-        {renderedResult}
+        <ToolResultDisplay
+          resultDisplay={resultDisplay}
+          availableTerminalHeight={availableTerminalHeight}
+          terminalWidth={terminalWidth}
+          renderOutputAsMarkdown={renderOutputAsMarkdown}
+        />
         {isThisShellFocused && config && (
           <Box paddingLeft={STATUS_INDICATOR_WIDTH} marginTop={1}>
             <ShellInputPrompt
@@ -279,98 +170,3 @@ export const ShellToolMessage: React.FC<ToolMessageProps> = ({
     </Box>
   );
 };
-
-type ToolStatusIndicatorProps = {
-  status: ToolCallStatus;
-  name: string;
-};
-
-const ToolStatusIndicator: React.FC<ToolStatusIndicatorProps> = ({
-  status,
-  name,
-}) => {
-  const isShell =
-    name === SHELL_COMMAND_NAME ||
-    name === SHELL_NAME ||
-    name === SHELL_TOOL_NAME;
-  const statusColor = isShell ? theme.ui.symbol : theme.status.warning;
-
-  return (
-    <Box minWidth={STATUS_INDICATOR_WIDTH}>
-      {status === ToolCallStatus.Pending && (
-        <Text color={theme.status.success}>{TOOL_STATUS.PENDING}</Text>
-      )}
-      {status === ToolCallStatus.Executing && (
-        <GeminiRespondingSpinner
-          spinnerType="toggle"
-          nonRespondingDisplay={TOOL_STATUS.EXECUTING}
-        />
-      )}
-      {status === ToolCallStatus.Success && (
-        <Text color={theme.status.success} aria-label={'Success:'}>
-          {TOOL_STATUS.SUCCESS}
-        </Text>
-      )}
-      {status === ToolCallStatus.Confirming && (
-        <Text color={statusColor} aria-label={'Confirming:'}>
-          {TOOL_STATUS.CONFIRMING}
-        </Text>
-      )}
-      {status === ToolCallStatus.Canceled && (
-        <Text color={statusColor} aria-label={'Canceled:'} bold>
-          {TOOL_STATUS.CANCELED}
-        </Text>
-      )}
-      {status === ToolCallStatus.Error && (
-        <Text color={theme.status.error} aria-label={'Error:'} bold>
-          {TOOL_STATUS.ERROR}
-        </Text>
-      )}
-    </Box>
-  );
-};
-
-type ToolInfo = {
-  name: string;
-  description: string;
-  status: ToolCallStatus;
-  emphasis: TextEmphasis;
-};
-const ToolInfo: React.FC<ToolInfo> = ({
-  name,
-  description,
-  status,
-  emphasis,
-}) => {
-  const nameColor = React.useMemo<string>(() => {
-    switch (emphasis) {
-      case 'high':
-        return theme.text.primary;
-      case 'medium':
-        return theme.text.primary;
-      case 'low':
-        return theme.text.secondary;
-      default: {
-        const exhaustiveCheck: never = emphasis;
-        return exhaustiveCheck;
-      }
-    }
-  }, [emphasis]);
-  return (
-    <Box overflow="hidden" height={1} flexGrow={1} flexShrink={1}>
-      <Text strikethrough={status === ToolCallStatus.Canceled} wrap="truncate">
-        <Text color={nameColor} bold>
-          {name}
-        </Text>{' '}
-        <Text color={theme.text.secondary}>{description}</Text>
-      </Text>
-    </Box>
-  );
-};
-
-const TrailingIndicator: React.FC = () => (
-  <Text color={theme.text.primary} wrap="truncate">
-    {' '}
-    ←
-  </Text>
-);
